Func: added sample/get_jobs.sh sample/get_projects.sh

+++ sample/get_jobs.sh
@@ -1,19 +1,30 @@
 #!/bin/sh
 
-DIR=output
-ITEM=IPAD
+while getopts i: var; do
+    case "$var" in
+    i) ITEM=$OPTARG ;;
+    esac
+done
+
+if [ -z "${ITEM}" ]; then
+    echo "Usage: get_jobs -i <ITEM>";
+    echo "( ITEM: IPHONE | IPAD | ANDROID | MACHINE-LEARNING | DATA-PROCESSING | DATA-MINING
+        NATURAL-LANGUAGE | ALGORITHM  | CRYTOGRAPHY )";
+    exit 1;
+fi
+
+DIR=jobs
 ITEM_PATH=${DIR}/${ITEM}
+TEMPLATE_LINK=https://www.freelancer.com/jobs/${ITEM}/XXX/
 
-TEMPLATE_LINK=https://www.freelancer.com/jobs/${ITEM}/XXX/index.html
-
-# get first index.htmal
+# get first index.html
 if [ -d "${ITEM_PATH}" ]; then
 	rm -rf ${ITEM_PATH}/*;
 else
-	mkdir ${ITEM_PATH};
+	mkdir -p ${ITEM_PATH};
 fi
 
-FIRST_LINK=https://www.freelancer.com/jobs/${ITEM}/1/index.html
+FIRST_LINK=https://www.freelancer.com/jobs/${ITEM}/1/
 echo "Downloading ${FIRST_LINK}";
 wget -O ${ITEM_PATH}/index.html ${FIRST_LINK} >& /dev/null;
 
@@ -37,14 +48,19 @@ echo "ALL:   ${ALL_ENTRIES}"
 echo "STEP:  ${ALL_PAGES}"
 echo "=========================="
 
-find_links ${ITEM_PATH}/index.html | grep "www.freelancer.com/projects";
-
 STEP=1
-#ALL_PAGES=5
+
+printf "ITEM: %d..." ${STEP};
+find_links ${ITEM_PATH}/index.html | grep "www.freelancer.com/projects" > ${DIR}/${ITEM}.link;
+
 while [ ${STEP} -lt ${ALL_PAGES} ]; do
     STEP=$(( ${STEP} + 1 ));
     LINK=`echo ${TEMPLATE_LINK} | sed -e "s/XXX/${STEP}/g"`;
-    #echo "Downloading ${LINK}";
+    printf "%d..." ${STEP};
     wget -O ${ITEM_PATH}/index_${STEP}.html ${LINK} >& /dev/null;
-    find_links ${ITEM_PATH}/index_${STEP}.html | grep "www.freelancer.com/projects";
+    find_links ${ITEM_PATH}/index_${STEP}.html | grep "www.freelancer.com/projects" >> ${DIR}/${ITEM}.link;
 done
+
+echo "FINSIHED";
+
+
